Use the same request processing loop for HTTP/1.x and HTTP/2.
