--- src/Kestrel.Core/Internal/Http/IMessageBody.cs
@@ -1,19 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
-{
-    public interface IMessageBody
-    {
-        bool RequestUpgrade { get; }
-
-        Task<int> ReadAsync(ArraySegment<byte> buffer, CancellationToken cancellationToken = default(CancellationToken));
-
-        Task CopyToAsync(Stream destination, CancellationToken cancellationToken = default(CancellationToken));
-    }
-}
