+++ test/Kestrel.Core.Tests/StreamsTests.cs
@@ -72,7 +72,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
             await upgrade.WriteAsync(new byte[1], 0, 1);
         }
 
-        private class MockMessageBody : MessageBody
+        private class MockMessageBody : Http1MessageBody
         {
             public MockMessageBody(bool upgradeable = false)
                 : base(null)
